requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
httpx
beautifulsoup4
pymysql
python-multipart
pandas
openpyxl
cryptography